+++ R/globals.R
@@ -1,8 +1,11 @@
-utils::globalVariables(c("questions_to_regrade", "student_identifier", 
-                         "assignment_missing", "feedback_path_Rmd",
-                         "feedback_path_to_be_knitted", "assignment_path",
-                         "grading_status", "feedback_codes", "graded_qs",
-                         "last_time_graded", 
-                         "students_in_team", "temp_grade_sheet_path",
-                         "points_to_remove", "student_identifier", 
-                         "team_identifier", "issue_titles"))
+utils::globalVariables(c(
+  "questions_to_regrade", "student_identifier", 
+  "assignment_missing", "feedback_path_Rmd",
+  "feedback_path_to_be_knitted", "assignment_path",
+  "grading_status", "feedback_codes", "graded_qs",
+  "last_time_graded", 
+  "students_in_team", "temp_grade_sheet_path",
+  "points_to_remove", "student_identifier", 
+  "team_identifier", "issue_titles",
+  "comments", "comment_qs"
+))