+++ R/assist_grading-wrappers.R
@@ -14,7 +14,7 @@
 #'     This file path structure will be used to determine where the other feedback files will be stored. 
 #'     The student identifier must be present somewhere in the file name and must be the only part of the file path unique to the student. 
 #'     The extension of the feedback file must be one of the following: "Rmd", "md", "docx", "html", "pdf". 
-#'     These file types (except the first) will be knitted to: Github, Word, html, and pdf documents respectively
+#'     These file types (except the first) will be knitted to: GitHub, Word, html, and pdf documents respectively
 #' @param example_student_identifier string; a student identifier (e.g. name, id, id number, GitHub user name) that is used to identify the student on the roster. 
 #'     This needs to be present somewhere in the example_assignment_path. The student_identifier needs to be the GitHub user name if the user wishes to push issues or feedback to GitHub later
 #' @param example_team_identifier string; Used instead of example_student_identifier when grading team assignments with assist_team_grading().