+++ NEWS.md
@@ -14,3 +14,4 @@
     - pdf
 - Moved `questions_to_grade` argument before `students_to_grade` in `assist_team_grading()`, and `assist_advanced_grading()`.
 - Moved `questions_to_regrade` argument before `students_to_regrade` in `assist_grading()`
+- temporary grade sheet is now called the grading progress log